+++ src/main/kotlin/de/mariushoefler/flutterenhancementsuite/models/RepositoryInfo.kt
@@ -1,0 +1,5 @@
+package de.mariushoefler.flutterenhancementsuite.models
+
+import com.google.gson.annotations.SerializedName
+
+data class RepositoryInfo(@SerializedName("default_branch") val defaultBranch: String)