+++ src/main/kotlin/de/mariushoefler/flutterenhancementsuite/utils/PubApi.kt
@@ -9,11 +9,13 @@ import de.mariushoefler.flutterenhancementsuite.exceptions.PubApiCouldNotBeReach
 import de.mariushoefler.flutterenhancementsuite.exceptions.PubApiUnknownFormat
 import de.mariushoefler.flutterenhancementsuite.models.PubPackage
 import de.mariushoefler.flutterenhancementsuite.models.PubPackageSearch
+import de.mariushoefler.flutterenhancementsuite.models.PubScore
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
+import retrofit2.create
 import retrofit2.http.GET
 import retrofit2.http.Path
 import java.io.IOException
@@ -30,12 +32,6 @@ object PubApi {
     }
 
     fun searchPackage(query: String, page: Int): PubPackageSearch? {
-        // +--- Uncomment to test bug report dialog
-        // |
-        // V
-        // val test = listOf(1,2,3)
-        // println(test[4])
-
         try {
             return runWithCheckCanceled {
                 val q = URLEncoder.encode(query, StandardCharsets.UTF_8.toString())
@@ -68,6 +64,19 @@ object PubApi {
         })
     }
 
+    @Throws(PubApiCouldNotBeReached::class)
+    fun getPackageScore(name: String, function: (response: Response<PubScore>) -> Unit) {
+        pubApiService.getPackageScore(name).enqueue(object : Callback<PubScore> {
+            override fun onResponse(call: Call<PubScore>, response: Response<PubScore>) {
+                return function(response)
+            }
+
+            override fun onFailure(call: Call<PubScore>, t: Throwable) {
+                throw PubApiCouldNotBeReached(Exception(t))
+            }
+        })
+    }
+
     @Throws(GetLatestPackageVersionException::class)
     fun getPackageLatestVersion(packageName: String): String {
         dependencyCache[packageName]?.let {
@@ -87,45 +96,48 @@ object PubApi {
     }
 
     fun getPackageChangelog(packageName: String): String? {
-        val pubPackage = lastPackages[packageName]
-            ?: pubApiService.getPackage(packageName).execute().body()
+        val pubPackage = lastPackages[packageName] ?: pubApiService.getPackage(packageName).execute().body()
 
-        if (pubPackage?.latest?.pubspec?.homepage == null) return null
+        if (pubPackage?.latest?.pubspec?.repository == null) return null
 
         val result = StringBuilder()
         result.append("<html>")
         result.append("<h1>$packageName Changelog</h1>")
 
-        val homepage = pubPackage.latest.pubspec.homepage
-        val src: String? = if (homepage.startsWith("https://github.com")) {
-            GithubApi.fetchContentsFromFile(homepage, "changelog")
+        val repository = pubPackage.latest.pubspec.repository
+        val src: String? = if (repository.startsWith("https://github.com")) {
+            GithubApi.fetchContentsFromFile(repository, "changelog")
         } else null
 
         return if (!src.isNullOrEmpty()) {
-            result.append(GithubApi.formatMarkdownAsHtml(src, pubPackage.latest.pubspec.homepage))
+            result.append(GithubApi.formatMarkdownAsHtml(src, pubPackage.latest.pubspec.repository))
             result.toString()
         } else null
     }
 
     fun getPackageDoc(packageName: String, short: Boolean = false): String? {
-        val pubPackage = lastPackages[packageName]
-            ?: pubApiService.getPackage(packageName).execute().body()
-            ?: return null
+        val pubPackage =
+            lastPackages[packageName] ?: pubApiService.getPackage(packageName).execute().body() ?: return null
 
         val result = StringBuilder()
         result.append("<html>")
         result.append("<h1>$packageName</h1>")
 
-        pubPackage.latest.pubspec.getAuthorName().let { authors ->
+        val pubspec = pubPackage.latest.pubspec
+        pubspec.getAuthorName().let { authors ->
             if (authors.isNotEmpty()) {
                 result.append("<small><i>by $authors</i></small><br><br>")
             }
         }
 
-        result.append("<p>${pubPackage.latest.pubspec.description}</p><br>")
+        result.append("<p>${pubspec.description}</p><br>")
 
-        if (!short && pubPackage.latest.pubspec.homepage != null) {
-            generateFullPackageDoc(pubPackage.latest.pubspec.homepage, result)
+        if (!short) {
+            if (pubspec.repository != null) {
+                generateFullPackageDoc(pubspec.repository, result)
+            } else if (pubspec.homepage != null) {
+                generateFullPackageDoc(pubspec.homepage, result)
+            }
         }
 
         result.append("</html>")
@@ -133,22 +145,21 @@ object PubApi {
         return result.toString()
     }
 
-    private fun generateFullPackageDoc(homepage: String, result: StringBuilder) {
-        val src: String? = if (homepage.startsWith("https://github.com")) {
-            GithubApi.fetchContentsFromFile(homepage, "readme")
+    private fun generateFullPackageDoc(repository: String, result: StringBuilder) {
+        val src: String? = if (repository.startsWith("https://github.com")) {
+            GithubApi.fetchContentsFromFile(repository, "readme")
         } else null
 
-        result.append("<a href=\"${homepage}\">Visit package's homepage</a><br><br>")
+        result.append("<a href=\"${repository}\">Visit package's homepage</a><br><br>")
 
-        if (homepage.startsWith("https://github.com")) {
-            val examplePath: String = if (homepage.contains("/tree/master")) {
+        if (repository.startsWith("https://github.com")) {
+            val examplePath: String = if (repository.contains("/tree/master")) {
                 ""
             } else {
                 "/tree/master"
             }
             result.append(
-                "<a href=\"${homepage}$examplePath/example\">" +
-                    "Show an example of how to use the package</a><br><br>"
+                "<a href=\"${repository}$examplePath/example\">" + "Show an example of how to use the package</a><br><br>"
             )
         }
 
@@ -165,7 +176,7 @@ object PubApi {
 
             result.append("<br><h2><u>Documentation</u></h2>")
 
-            result.append(GithubApi.formatMarkdownAsHtml(html, homepage))
+            result.append(GithubApi.formatMarkdownAsHtml(html, repository))
         }
     }
 }
@@ -174,13 +185,15 @@ private interface PubApiService {
     @GET("packages/{name}")
     fun getPackage(@Path("name") name: String): Call<PubPackage>
 
+    @GET("packages/{name}/score")
+    fun getPackageScore(@Path("name") name: String): Call<PubScore>
+
     companion object {
         fun create(): PubApiService {
             val retrofit =
                 Retrofit.Builder().addConverterFactory(GsonConverterFactory.create()).baseUrl("https://pub.dev/api/")
                     .build()
-
-            return retrofit.create(PubApiService::class.java)
+            return retrofit.create<PubApiService>()
         }
     }
 }